requirements.txt
bcrypt==4.2.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
unittest-xml-reporting==3.0.4